Code clean up for parameterized models
Refactorings and code clean up folloiwing commit "Introduce creation of
parameterized tooling models"
