+++ subprojects/tooling-api-builders/src/test/groovy/org/gradle/tooling/internal/provider/runner/DefaultBuildControllerTest.groovy
@@ -21,24 +21,20 @@ import org.gradle.api.internal.GradleInternal
 import org.gradle.api.internal.project.ProjectInternal
 import org.gradle.initialization.BuildCancellationToken
 import org.gradle.internal.service.ServiceRegistry
-import org.gradle.tooling.internal.adapter.ProtocolToModelAdapter
-import org.gradle.tooling.internal.adapter.ViewBuilder
 import org.gradle.tooling.internal.gradle.GradleProjectIdentity
 import org.gradle.tooling.internal.protocol.InternalUnsupportedModelException
 import org.gradle.tooling.internal.protocol.ModelIdentifier
 import org.gradle.tooling.provider.model.ToolingModelBuilder
 import org.gradle.tooling.provider.model.ToolingModelBuilderRegistry
-import org.gradle.tooling.provider.model.ToolingParameterizedModelBuilder
+import org.gradle.tooling.provider.model.ParameterizedToolingModelBuilder
 import org.gradle.tooling.provider.model.UnknownModelException
 import spock.lang.Specification
 
 class DefaultBuildControllerTest extends Specification {
     def cancellationToken = Stub(BuildCancellationToken)
-    def adapter = Stub(ProtocolToModelAdapter)
     def gradle = Stub(GradleInternal) {
         getServices() >> Stub(ServiceRegistry) {
             get(BuildCancellationToken) >> cancellationToken
-            get(ProtocolToModelAdapter) >> adapter
         }
     }
     def registry = Stub(ToolingModelBuilderRegistry)
@@ -51,7 +47,7 @@ class DefaultBuildControllerTest extends Specification {
         getName() >> 'some.model'
     }
     def modelBuilder = Stub(ToolingModelBuilder)
-    def parameterizedModelBuilder = Stub(ToolingParameterizedModelBuilder)
+    def parameterizedModelBuilder = Stub(ParameterizedToolingModelBuilder)
     def controller = new DefaultBuildController(gradle)
 
     def "adapts model not found exception to protocol exception"() {
@@ -136,18 +132,25 @@ class DefaultBuildControllerTest extends Specification {
     def "uses parameterized builder when parameter is not null"() {
         def model = new Object()
         def wrongModel = new Object()
-        def parameterType = Object.class
-        def viewBuilder = Stub(ViewBuilder)
-        def parameter = new Object()
-        def internalParameter = new Object()
+        def parameterType = CustomParameter.class
+        def parameter = new CustomParameter() {
+            @Override
+            String getValue() {
+                return "myValue"
+            }
+            @Override
+            void setValue(String value) {}
+        }
 
         given:
         _ * gradle.defaultProject >> project
         _ * registry.getBuilder("some.model") >> parameterizedModelBuilder
         _ * parameterizedModelBuilder.getParameterType() >> parameterType
-        _ * adapter.builder(parameterType) >> viewBuilder
-        _ * viewBuilder.build(parameter) >> internalParameter
-        _ * parameterizedModelBuilder.buildAll("some.model", internalParameter, project) >> model
+        _ * parameterizedModelBuilder.buildAll("some.model", _, project) >> { def modelName, CustomParameter param, projectInternal ->
+            assert param != null
+            assert param.getValue() == "myValue"
+            return model
+        }
         _ * parameterizedModelBuilder.buildAll("some.model", project) >> wrongModel
 
         when:
@@ -166,9 +169,14 @@ class DefaultBuildControllerTest extends Specification {
         _ * modelBuilder.buildAll("some.model", project) >> model
 
         when:
-        def result = controller.getModel(null, modelId, new Object())
+        controller.getModel(null, modelId, new Object())
 
         then:
         thrown(InternalUnsupportedModelException)
+    }
+
+    interface CustomParameter {
+        String getValue()
+        void setValue(String value)
     }
 }
