+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/internal/consumer/connection/TestExecutionConsumerConnection.java
@@ -24,7 +24,7 @@ import org.gradle.tooling.internal.protocol.ConnectionVersion4;
 import org.gradle.tooling.internal.protocol.test.InternalTestExecutionConnection;
 import org.gradle.tooling.internal.consumer.TestExecutionRequest;
 
-/*
+/**
  * <p>Used for providers >= 2.6.</p>
  */
 public class TestExecutionConsumerConnection extends ShutdownAwareConsumerConnection {
